Created Level Set Methods and persistent variables
Now you can use methods in the Game Manager to set a specific levels gold and star goals. It also contains the functionality to know when a level starts and ends, and set global variables to stay the same between levels. Between startup doesn't work yet, but that's a whole different beast that will probably take me a week or two just on its own.